Renamed allotBatch to batch
