+++ dashboard/data.php
@@ -9,7 +9,7 @@ if(!isset($_SESSION['userId'])){
 	<button title="students" onclick="loadContent(this.title)">Student Detail</button>
 	<!-- <button title="pendingStudents" onclick="loadContent(this.title)">Pending </button>
 	<button title="completedStudents" onclick="loadContent(this.title)">Completed </button> -->
-	<button title="allotBatch" onclick="loadContent(this.title)">Batch Details</button>
+	<button title="batch" onclick="loadContent(this.title)">Batch Details</button>
 	<button title="otherOptions" onclick="loadContent(this.title)">Status Details</button>
 	<button title="coordinators" onclick="loadContent(this.title)">Co-Ordinators</button>
 	<button title="otherOptions" onclick="loadContent(this.title)">Other Options</button>
